refactoring: renames/add folder  and add alias path
