+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from "next/app";
 import { SessionProvider } from "next-auth/react";
-import "../scss/Global/reset.scss";
+import "@styles/Global/reset.scss";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
