+++ pages/api/me.ts
@@ -1,6 +1,8 @@
-import { getSpotifyMe } from "../../functions";
 import { getSession } from "next-auth/react";
+
 import type { NextApiRequest, NextApiResponse } from "next";
+
+import { getSpotifyMe } from "@providers/spotify";
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const session = await getSession({ req });
